refactor, componentize, cleanup
