+++ src/viewer/Viewer.js
@@ -1,6 +1,3 @@
-// CHANGE #14
-// remove keyboard input
-
 import * as Croquet from "@croquet/croquet";
 import ViewerModel from "./ViewerModel";
 import ViewerView from "./ViewerView";
@@ -12,8 +9,7 @@ const Viewer = () => {
   const pmaEventHandler = new PMAEventHandler();
   const xrpkAppId = pmaEventHandler.getAppState().appId;
   const appId = "com.plutovr.threesixtyviewer";
-  // const name = xrpkAppId ? xrpkAppId : Math.floor(Math.random() * 123456789);
-  const name = xrpkAppId ? xrpkAppId : "test9";
+  const name = xrpkAppId ? xrpkAppId : Math.floor(Math.random() * 123456789);
   const password = "password";
 
   Croquet.Session.join({
